Restructuring of layout to better fit gem convention.
