--- lib/config/adsr.rb
@@ -1,3 +1,0 @@
-def default(input)
-  Math.cos(2* Math::PI*input).to_f
-end
