+++ lib/muse/config.rb
@@ -1,0 +1,1 @@
+Dir["#{File.dirname(__FILE__)}/config/*.rb"].each{|f| require f }
