+++ lib/muse/config/adsr.rb
@@ -1,0 +1,9 @@
+module Muse
+  module Config
+    class << self
+      def default(input)
+        Math.cos(2* Math::PI*input).to_f
+      end
+    end
+  end
+end
