+++ songs/music_box_dancer.rb
@@ -1,4 +1,5 @@
 require "#{File.dirname(__FILE__)}/../lib/muse"
+include Muse
 
 # "Music Box Dancer" is an instrumental piece by Canadian musician 
 # Frank Mills that was an international hit in the late 1970s. It
