+++ songs/test.rb
@@ -1,4 +1,5 @@
 require "#{File.dirname(__FILE__)}/../lib/muse"
+include Muse
 
 Song.record 'test_file' do
   bar(1,b:0.5).notes { a; b b:1;}
